Added debugging context methods.
It is a hassle to always specify the debugging context within functions
since these usually not change in a small context. Now it is possible to
specify the debugging context on top of a function and only state the
debugging message.
Note that the context is NOT resetted at the end of the function. This
is not possible in easy (and clean) ways.